updated compats and started setting up REVISE generator
