+++ Project.toml
@@ -1,13 +1,12 @@
 name = "CounterfactualExplanations"
 uuid = "2f13d31b-18db-44c1-bc43-ebaf2cff0be0"
-authors = ["Patrick Altmeyer"]
+authors = ["Patrick Altmeyer <p.altmeyer@tudelft.nl>"]
 version = "0.1.1"
 
 [deps]
 BSON = "fbb218c0-5317-5bc6-957e-2ee96dd4b1f0"
 Flux = "587475ba-b771-5e3f-ad9e-33799f191a9c"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
-NNlib = "872c559c-99b0-510c-b3b7-b6c96a88d5cd"
 Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 PyCall = "438e738f-606a-5dbb-bf0a-cddfbfd45ab0"
@@ -18,9 +17,10 @@ StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 
 [compat]
 BSON = "0.3"
-Flux = "0.12"
-NNlib = "0.8"
+Flux = "0.12, 0.13"
 Plots = "1"
+PyCall = "1"
+RCall = "0.13"
 StatsBase = "0.33"
 julia = "1.5, 1.6, 1.7"
