+++ src/generators/gradient_based/GenericGenerator.jl
@@ -1,0 +1,36 @@
+using LinearAlgebra
+
+# -------- Wachter et al (2018): 
+struct GenericGenerator <: AbstractGradientBasedGenerator
+    loss::Symbol # loss function
+    complexity::Function # complexity function
+    λ::AbstractFloat # strength of penalty
+    ϵ::AbstractFloat # step size
+    τ::AbstractFloat # tolerance for convergence
+end
+
+"""
+    GenericGenerator(
+        ;
+        loss::Symbol=:logitbinarycrossentropy,
+        complexity::Function=norm,
+        λ::AbstractFloat=0.1,
+        ϵ::AbstractFloat=0.1,
+        τ::AbstractFloat=1e-5
+    )
+
+An outer constructor method that instantiates a generic generator.
+
+# Examples
+```julia-repl
+generator = GenericGenerator()
+```
+"""
+GenericGenerator(
+    ;
+    loss::Symbol=:logitbinarycrossentropy,
+    complexity::Function=norm,
+    λ::AbstractFloat=0.1,
+    ϵ::AbstractFloat=0.1,
+    τ::AbstractFloat=1e-5
+) = GenericGenerator(loss, complexity, λ, ϵ, τ)
